--- spamanalyzer/analyzer/__init__.py
@@ -1,23 +1,0 @@
-from spamanalyzer.analyzer.data_structures import (
-    Domain,
-    MailAnalysis,
-    MailAnalyzer,
-    Date,
-)
-from spamanalyzer.analyzer.utils import (
-    get_domain,
-    inspect_headers,
-    inspect_body,
-    inspect_attachments,
-)
-
-__all__ = [
-    "Domain",
-    "MailAnalysis",
-    "MailAnalyzer",
-    "get_domain",
-    "inspect_attachments",
-    "inspect_body",
-    "inspect_headers",
-    "Date",
-]
