--- spamanalyzer/cli/__init__.py
@@ -1,3 +1,0 @@
-from spamanalyzer.cli.run import main
-
-__all__ = ["main"]
